refactor: 구매 내역 페이지의 API 데이터 로드 및 상태 관리 개선, UI 레이아웃 조정
